+++ lib/pcanvas.dart
@@ -2,3 +2,4 @@
 library pcanvas;
 
 export 'src/pcanvas_base.dart';
+export 'src/pcanvas_element.dart';