+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.2
+
+- `PCanvasEvent` now is abstract:
+  - `PCanvasClickEvent`: for mouse and touch events.
+  - `PCanvasKeyEvent`: for keyboatd events.
+- `PCanvasPainter`:
+  - Added support to key events
+- `PCanvas`:
+  - Added support to fill gradient operations.
+- Added `PCanvasElement`
+  - Base class for personalized canvas elements. 
+
 ## 1.0.1
 
 - Added interface `WithDimension` to `PDimension` and `PCanvas`.